removes unneeded files from /NFT directory and re-organized file structure
