--- src/pages/NFTs/Analytics/mockData.js
@@ -1,180 +1,0 @@
-export const mockAnalyticsDrodown = [
-  { name: 'High to Low', icon: 'all' },
-  { name: 'Low to High', icon: 'art' }
-]
-
-export const mockFloorTabData = [
-  {
-    id: '1',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '2',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '3',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '4',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '5',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '6',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '7',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '8',
-    title: 'Degen Ape...',
-    value: '1.96 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  }
-]
-
-export const mockVolumeTabData = [
-  {
-    id: '1',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '2',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '3',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '4',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '5',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '6',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '7',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '8',
-    title: 'Corrupt Catz',
-    value: '2.05 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  }
-]
-export const mockListNFTTabData = [
-  {
-    id: '1',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '2',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '3',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '4',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '5',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '6',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '7',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'increase-icon',
-    percent: '+ 1.15 %'
-  },
-  {
-    id: '8',
-    title: 'A basic ape',
-    value: '1.70 SOL',
-    progress: 'decrease-icon',
-    percent: '+ 1.15 %'
-  }
-]
