--- src/pages/NFTs/Collection/mockData.js
@@ -1,84 +1,0 @@
-export const NFTsData = [
-  {
-    id: '1',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 1499,
-    status: 'sold_out',
-    hearts: 0,
-    remaining: '02d:20h:10min',
-    isFeatured: false,
-    isFavorite: false
-  },
-  {
-    id: '2',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 10000,
-    status: 'auctioning',
-    hearts: 0,
-    remaining: '02d:20h:10min',
-    isFeatured: true,
-    isFavorite: false
-  },
-  {
-    id: '3',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 420,
-    status: 'auctioning',
-    hearts: 8,
-    remaining: '02d:20h:10min',
-    isFeatured: false,
-    isFavorite: true
-  },
-  {
-    id: '4',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 420,
-    status: 'auctioning',
-    hearts: 14,
-    remaining: '02d:20h:10min',
-    isFeatured: false,
-    isFavorite: false
-  },
-  {
-    id: '5',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 2500,
-    status: 'auctioning',
-    hearts: 14,
-    remaining: '02d:20h:10min',
-    isFeatured: false,
-    isFavorite: true
-  },
-  {
-    id: '6',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 200000,
-    status: 'auctioning',
-    hearts: 14,
-    remaining: '02d:20h:10min',
-    isFeatured: false,
-    isFavorite: false
-  },
-  {
-    id: '7',
-    thumbnail: 'https://placeimg.com/30/30',
-    name: 'Genesis #3504',
-    price: 220000,
-    status: 'auctioning',
-    hearts: 14,
-    remaining: '02d:20h:10min',
-    isFeatured: false,
-    isFavorite: false
-  }
-]
-
-export const mockAnalyticsDrodown = [
-  { name: 'High to Low', icon: 'all' },
-  { name: 'Low to High', icon: 'art' }
-]
