--- src/pages/NFTs/CreateNFT/index.ts
@@ -1,1 +1,0 @@
-export * from './MyCreatedNFT'
