--- src/pages/NFTs/CreateNFT/mockData.js
@@ -1,195 +1,0 @@
-export const myCreatedNFTData = {
-  image: 'https://placeimg.com/465/465',
-  title: 'Auction ends in:',
-  likedCount: '7',
-  remaining: {
-    days: '03',
-    hours: '11',
-    minutes: '20'
-  },
-  type: 'Created By You',
-  price: '69 SOL',
-  fiat: '(10,902 USD aprox)',
-  percent: '+1.15%',
-  subTitle: 'Ethernal #03',
-  description: 'Etheranl is a collection of 1/1 beautifull abstract pieces taken from the mind Jackson Pollock.',
-  creator: {
-    image: 'https://placeimg.com/30/30',
-    name: 'yeoohr'
-  },
-  collection: {
-    image: 'https://placeimg.com/30/30',
-    name: 'Ethernal'
-  },
-  category: {
-    image: 'https://placeimg.com/30/30',
-    name: 'Art'
-  },
-  table: [
-    {
-      title: 'Mint address',
-      content: 'uHT2....4s6'
-    },
-    {
-      title: 'Token address',
-      content: 'QWRA2a...JK5'
-    },
-    {
-      title: 'Owner',
-      content: 'CY5ui...WE4'
-    },
-    {
-      title: 'Artist Royalties',
-      content: '5.5%'
-    },
-    {
-      title: 'Transaction Fee',
-      content: '2%'
-    }
-  ]
-}
-
-export const myOwnNFTData = {
-  image: 'https://placeimg.com/465/465',
-  title: 'Auction ends in:',
-  likedCount: '7',
-  remaining: {
-    days: '06',
-    hours: '11',
-    minutes: '04'
-  },
-  type: 'Owned By You',
-  price: '69 SOL',
-  fiat: '(10,902 USD aprox)',
-  percent: '+3.22%',
-  subTitle: '#1695',
-  description: 'Corrupt Catz hail from all over the globe - a special species of 5,684 felines. ',
-  creator: {
-    image: 'https://placeimg.com/30/30',
-    name: 'Corrupt...'
-  },
-  collection: {
-    image: 'https://placeimg.com/30/30',
-    name: 'The lions'
-  },
-  category: {
-    image: 'https://placeimg.com/30/30',
-    name: 'Art'
-  },
-  table: [
-    {
-      title: 'Mint address',
-      content: 'uHT2....4s6'
-    },
-    {
-      title: 'Token address',
-      content: 'QWRA2a...JK5'
-    },
-    {
-      title: 'Owner',
-      content: 'CY5ui...WE4'
-    },
-    {
-      title: 'Artist Royalties',
-      content: '5.5%'
-    },
-    {
-      title: 'Transaction Fee',
-      content: '2%'
-    },
-    {
-      title: 'Fee',
-      content: '2%'
-    }
-  ]
-}
-
-export const detailTabData = [
-  {
-    title: 'Mint address',
-    value: 'uHT2....4s6'
-  },
-  {
-    title: 'Token address',
-    value: 'QWRA2a...JK5'
-  },
-  {
-    title: 'Owner',
-    value: 'CY5ui...WE4'
-  },
-  {
-    title: 'Artist Royalties',
-    value: '5.5%'
-  },
-  {
-    title: 'Transaction Fee',
-    value: '2%'
-  }
-]
-
-export const tradingHistoryTabData = [
-  {
-    id: '1',
-    event: 'list',
-    price: 150,
-    from: 'Evan34',
-    to: '',
-    date: '11/10/21'
-  },
-  {
-    id: '2',
-    event: 'offer',
-    price: 120.5678,
-    from: 'capital_1',
-    to: '',
-    date: '09/10/21'
-  },
-  {
-    id: '3',
-    event: 'offer',
-    price: 135.556,
-    from: 'MLBmodel',
-    to: 'Chirsstoo',
-    date: '02/10/21'
-  },
-  {
-    id: '4',
-    event: 'sale',
-    price: 121.134,
-    from: 'Chirsstoo',
-    to: '',
-    date: '25/09/21'
-  },
-  {
-    id: '5',
-    event: 'list',
-    price: 150,
-    from: 'Evan34',
-    to: 'Chirsstoo',
-    date: '11/10/21'
-  },
-  {
-    id: '6',
-    event: 'offer',
-    price: 120.5678,
-    from: 'capital_1',
-    to: '',
-    date: '09/10/21'
-  },
-  {
-    id: '7',
-    event: 'offer',
-    price: 135.556,
-    from: 'MLBmodel',
-    to: '',
-    date: '02/10/21'
-  },
-  {
-    id: '8',
-    event: 'sale',
-    price: 121.134,
-    from: 'Chirsstoo',
-    to: '',
-    date: '25/09/21'
-  }
-]
