--- src/pages/NFTs/FixedPriceNFT/mockData.js
@@ -1,109 +1,0 @@
-export const NFTDetailsGeneralData = {
-  name: 'Genesis #3886',
-  image: `/img/assets/card-1.png`,
-  isFavorite: true,
-  hearts: 6,
-  remaining: {
-    days: '04',
-    hours: '30',
-    minutes: '35'
-  },
-  price: 150000,
-  fiat: '21,900 USD aprox',
-  isForCharity: true,
-  percent: '+ 1.15 %',
-  intro: '9,999 levitating 3D cities ready to expand, conquer and defend into the metaverse',
-  creator: {
-    title: 'Solcities',
-    thumbnail: `/img/assets/avatar@3x.png`
-  },
-  collection: {
-    title: 'Solcities',
-    thumbnail: `/img/assets/avatar@3x.png`
-  },
-  category: {
-    title: 'Metaverse',
-    thumbnail: `/img/assets/avatar@3x.png`
-  }
-}
-
-export const detailTabData = [
-  {
-    title: 'Mint address',
-    value: 'uHT2....4s6'
-  },
-  {
-    title: 'Token address',
-    value: 'QWRA2a...JK5'
-  },
-  {
-    title: 'Owner',
-    value: 'CY5ui...WE4'
-  },
-  {
-    title: 'Artist Royalties',
-    value: '5.5%'
-  },
-  {
-    title: 'Transaction Fee',
-    value: '2%'
-  }
-]
-
-export const tradingHistoryTabData = [
-  {
-    id: '1',
-    event: 'list',
-    price: 150,
-    from: 'Evan34',
-    to: '',
-    date: '11/10/21'
-  },
-  {
-    id: '2',
-    event: 'offer',
-    price: 120.5678,
-    from: 'capital_1',
-    to: '',
-    date: '09/10/21'
-  },
-  {
-    id: '3',
-    event: 'offer',
-    price: 135.556,
-    from: 'MLBmodel',
-    to: 'Chirsstoo',
-    date: '02/10/21'
-  },
-  {
-    id: '4',
-    event: 'sale',
-    price: 121.134,
-    from: 'Chirsstoo',
-    to: '',
-    date: '25/09/21'
-  }
-]
-
-export const attributesTabContentData = [
-  {
-    title: 'Atmpsphere',
-    value: 'Ocen Spary'
-  },
-  {
-    title: 'Underground',
-    value: 'Night Moonstones'
-  },
-  {
-    title: 'Land',
-    value: 'Glass Atlantis'
-  },
-  {
-    title: 'Resources',
-    value: 'Minerals'
-  },
-  {
-    title: 'Transaction Fee',
-    value: '2%'
-  }
-]
