+++ src/pages/NFTs/Home/CollectionCarousel.tsx
@@ -3,11 +3,11 @@ import Slider from 'react-slick'
 import 'slick-carousel/slick/slick.css'
 import 'slick-carousel/slick/slick-theme.css'
 import styled from 'styled-components'
-import { ArrowClicker, Loader } from '../../components'
-import { ButtonWrapper } from './NFTButton'
+import { ArrowClicker, Loader } from '../../../components'
+import { ButtonWrapper } from '../NFTButton'
 import NFTImageCarouselItem from './NFTImageCarouselItem'
-import { NFTBaseCollection, NFTFeaturedCollection, NFTUpcomingCollection } from '../../types/nft_collections.d'
-import { SkeletonCommon } from './Skeleton/SkeletonCommon'
+import { NFTBaseCollection, NFTFeaturedCollection, NFTUpcomingCollection } from '../../../types/nft_collections.d'
+import { SkeletonCommon } from '../Skeleton/SkeletonCommon'
 import isEmpty from 'lodash/isEmpty'
 
 const CAROUSEL_WRAPPER = styled.div`
