+++ src/pages/NFTs/Home/Header.tsx
@@ -8,7 +8,7 @@ import { ButtonWrapper } from '../NFTButton'
 // import { SearchBar, Categories } from '../../../components'
 import { SearchBar } from '../../../components'
 import { SpaceBetweenDiv } from '../../../styles'
-// import { categories, coins } from './mockData'
+// import { categories, coins } from './data'
 // import { useLocalStorageState } from '../../../utils'
 import PopupCompleteProfile from '../Profile/PopupCompleteProfile'
 import { useNFTProfile } from '../../../context'
