+++ src/pages/NFTs/Home/NFTHome.tsx
@@ -1,11 +1,11 @@
 import React, { useRef, useState, useEffect, FC } from 'react'
 import styled from 'styled-components'
-import { Header } from '../Header'
+import { Header } from './Header'
 import { NFTHomeSlider } from '../Slider/NFTHomeSlider'
-import AnalyticsTabs from '../Analytics/Tab'
-import Loading from '../Analytics/Loading'
-import NFTFooter from '../NFTFooter'
-import CollectionCarousel from '../CollectionCarousel'
+import AnalyticsTabs from './Tab'
+import Loading from './Loading'
+import NFTFooter from './NFTFooter'
+import CollectionCarousel from './CollectionCarousel'
 import { useNFTCollections } from '../../../context'
 import { COLLECTION_TYPES } from '../../../types/nft_collections.d'
 import { SVGDynamicReverseMode } from '../../../styles'
