--- src/pages/NFTs/Home/mockData.js
@@ -1,70 +1,0 @@
-export const allCollections = [
-  {
-    collection_id: 1,
-    collection_name: 'Thugbirdz',
-    collection_description: 'A collection of 3333 nft birds on solana',
-    profile_pic_link:
-      'https://images.unsplash.com/photo-1634985492257-06c8ee26b770?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1932&q=80',
-    banner_link: 'RANDOM_LINK',
-    banner_2_link: null,
-    banner_3_link: null,
-    title: 'Thugbirdz',
-    tagline: 'Birds that are thugs, yo',
-    size: 3333,
-    category_tags: 'birds thugs art pfp',
-    is_verified: true
-  },
-  {
-    collection_id: 2,
-    collection_name: 'Degen apes',
-    collection_description: 'A collection of 5000 ape pfps on Solana',
-    profile_pic_link:
-      'https://images.unsplash.com/photo-1634985492257-06c8ee26b770?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1932&q=80',
-    banner_link: 'RANDOM_LINK',
-    banner_2_link: null,
-    banner_3_link: null,
-    title: 'Degen apes',
-    tagline: '5000 degens aping on solana',
-    size: 3333,
-    category_tags: 'apes degen art pfp',
-    is_verified: true
-  }
-]
-
-export const featuredCollections = [
-  {
-    featured_collection_name: 'SuperFrogs',
-    featured_collection_banner_url:
-      'https://images.unsplash.com/photo-1634985492257-06c8ee26b770?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1932&q=80',
-    featured_collection_description:
-      'A collection of super frogs on the solana blockchain, only 9999 frogs exist! Catch them all!',
-    collection_id: 1
-  },
-  {
-    featured_collection_name: 'Cyber Horses',
-    featured_collection_banner_url:
-      'https://images.unsplash.com/photo-1634985492257-06c8ee26b770?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1932&q=80',
-    featured_collection_description: '7777 cyberpunk horses on the Solana ranchchain',
-    collection_id: 2
-  }
-]
-
-export const upcomingCollections = [
-  {
-    upcoming_collection_droptime: 'SOME_UNIX_TIMESTAMP',
-    upcoming_collection_name: 'SuperFrogs',
-    upcoming_collection_banner_url:
-      'https://images.unsplash.com/photo-1634985492257-06c8ee26b770?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1932&q=80',
-    upcoming_collection_description:
-      'A collection of super frogs on the solana blockchain, only 9999 frogs exist! Catch them all!',
-    upcoming_id: null
-  },
-  {
-    upcoming_collection_droptime: 'SOME_UNIX_TIMESTAMP',
-    upcoming_collection_name: 'Cyber Horses',
-    upcoming_collection_banner_url:
-      'https://images.unsplash.com/photo-1634985492257-06c8ee26b770?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1932&q=80',
-    upcoming_collection_description: '7777 cyberpunk horses on the Solana ranchchain',
-    collection_id: 2
-  }
-]
