+++ src/pages/NFTs/NFTDetails/BidModal.tsx
