--- src/pages/NFTs/OpenBidNFT/index.ts
@@ -1,1 +1,0 @@
-export * from './OpenBidNFT'
