--- src/pages/NFTs/Profile/mockData.js
@@ -1,80 +1,0 @@
-export const mockData = [
-  {
-    event: 'Transfer',
-    item: {
-      image: '',
-      name: 'Oceans #2',
-      other: 'capital_1'
-    },
-    price: '35.5 SOL',
-    quantity: '1',
-    from: 'yeoohr',
-    to: 'Evan34',
-    date: '09/10/21'
-  },
-  {
-    event: 'Offer',
-    item: {
-      image: '',
-      name: 'Oceans #2',
-      other: 'capital_1'
-    },
-    price: '35.5 SOL',
-    quantity: '1',
-    from: 'yeoohr',
-    to: 'Evan34',
-    date: '09/10/21'
-  },
-  {
-    event: 'List',
-    item: {
-      image: '',
-      name: 'Oceans #2',
-      other: 'capital_1'
-    },
-    price: '35.5 SOL',
-    quantity: '1',
-    from: 'yeoohr',
-    to: 'Evan34',
-    date: '09/10/21'
-  },
-  {
-    event: 'Sale',
-    item: {
-      image: '',
-      name: 'Oceans #2',
-      other: 'capital_1'
-    },
-    price: '35.5 SOL',
-    quantity: '1',
-    from: 'yeoohr',
-    to: 'Evan34',
-    date: '09/10/21'
-  },
-  {
-    event: 'Minted',
-    item: {
-      image: '',
-      name: 'Oceans #2',
-      other: 'capital_1'
-    },
-    price: '35.5 SOL',
-    quantity: '1',
-    from: 'yeoohr',
-    to: 'Evan34',
-    date: '09/10/21'
-  },
-  {
-    event: 'Oceans',
-    item: {
-      image: '',
-      name: 'Oceans #2',
-      other: 'capital_1'
-    },
-    price: '35.5 SOL',
-    quantity: '1',
-    from: 'yeoohr',
-    to: 'Evan34',
-    date: '09/10/21'
-  }
-]
